# blockchain_tools/api/openapi/blockchain.yaml
openapi: 3.0.0
info:
  title: Blockchain API
  description: API for interacting with the blockchain
  version: 1.0.0
  contact:
    name: Layla Abdullah
    url: https://github.com/layla-lili/blockchain_mvp

servers:
  - url: /api/v1
    description: Development server

paths:
  /blocks/{number}:
    get:
      summary: Get block by number
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Block details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"

  /transactions:
    post:
      summary: Send a new transaction
      operationId: sendTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "200":
          description: Transaction sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Invalid transaction parameters
    get:
      summary: List recent transactions
      operationId: listTransactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  /transactions/{hash}:
    get:
      summary: Get transaction details
      operationId: getTransaction
      parameters:
        - name: hash
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /accounts/{address}/balance:
    get:
      summary: Get account balance
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"

components:
  schemas:
    Block:
      type: object
      properties:
        hash:
          type: string
        number:
          type: integer
        timestamp:
          type: integer
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"

    Transaction:
      type: object
      properties:
        hash:
          type: string
        from:
          type: string
        to:
          type: string
        value:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
        blockNumber:
          type: integer
        timestamp:
          type: integer

    TransactionRequest:
      type: object
      required:
        - to
        - value
      properties:
        from:
          type: string
          description: Sender address (optional for test transactions)
        to:
          type: string
          description: Recipient address
        value:
          type: string
          description: Transaction value in wei
        data:
          type: string
          description: Optional transaction data

    TransactionResponse:
      type: object
      properties:
        hash:
          type: string
          description: Transaction hash
        status:
          type: string
          enum: [pending, confirmed, failed]

    Balance:
      type: object
      properties:
        address:
          type: string
        amount:
          type: string